+++ tests/CppUTestExt/MockSupportTest.cpp
@@ -44,79 +44,6 @@ TEST_GROUP(MockSupportTest)
   }
 };
 
-class MyTypeForTesting
-{
-public:
-    MyTypeForTesting(long val)
-    {
-        value = new long(val);
-    }
-    virtual ~MyTypeForTesting()
-    {
-        delete value;
-    }
-    long *value;
-};
-
-class MyTypeForTestingComparator : public MockNamedValueComparator
-{
-public:
-    virtual bool isEqual(const void* object1, const void* object2)
-    {
-        const MyTypeForTesting* obj1 = (const MyTypeForTesting*) object1;
-        const MyTypeForTesting* obj2 = (const MyTypeForTesting*) object2;
-        return *(obj1->value) == *(obj2->value);
-    }
-    virtual SimpleString valueToString(const void* object)
-    {
-        const MyTypeForTesting* obj = (const MyTypeForTesting*) object;
-        return StringFrom(*(obj->value));
-    }
-};
-
-class MyTypeForTestingCopier : public MockNamedValueCopier
-{
-public:
-    virtual void copy(void* dst_, const void* src_)
-    {
-        MyTypeForTesting* dst = (MyTypeForTesting*) dst_;
-        const MyTypeForTesting* src = (const MyTypeForTesting*) src_;
-        *(dst->value) = *(src->value);
-    }
-};
-
-TEST(MockSupportTest, customObjectParameterFailsWhenNotHavingAComparisonRepository)
-{
-    MyTypeForTesting object(1);
-    mock().expectOneCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-    mock().actualCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-
-    MockNoWayToCompareCustomTypeFailure expectedFailure(mockFailureTest(), "MyTypeForTesting");
-    CHECK_EXPECTED_MOCK_FAILURE(expectedFailure);
-}
-
-TEST(MockSupportTest, customObjectParameterFailsWhenNotHavingACopierRepository)
-{
-    MyTypeForTesting object(1);
-    mock().expectOneCall("function").withOutputParameterOfTypeReturning("MyTypeForTesting", "parameterName", &object);
-    mock().actualCall("function").withOutputParameterOfType("MyTypeForTesting", "parameterName", &object);
-
-    MockNoWayToCopyCustomTypeFailure expectedFailure(mockFailureTest(), "MyTypeForTesting");
-    CHECK_EXPECTED_MOCK_FAILURE(expectedFailure);
-}
-
-TEST(MockSupportTest, customObjectParameterSucceeds)
-{
-    MyTypeForTesting object(1);
-    MyTypeForTestingComparator comparator;
-    mock().installComparator("MyTypeForTesting", comparator);
-    mock().expectOneCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-    mock().actualCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-    mock().checkExpectations();
-    CHECK_NO_MOCK_FAILURE();
-    mock().removeAllComparatorsAndCopiers();
-}
-
 TEST(MockSupportTest, outputParameterSucceeds)
 {
     int param = 1;
@@ -306,363 +233,6 @@ TEST(MockSupportTest, outputParameterWithIgnoredParameters)
 
     LONGS_EQUAL(param, retval);
 }
-
-static bool myTypeIsEqual(const void* object1, const void* object2)
-{
-    return ((MyTypeForTesting*)object1)->value == ((MyTypeForTesting*)object2)->value;
-}
-
-static SimpleString myTypeValueToString(const void* object)
-{
-    return StringFrom(((MyTypeForTesting*)object)->value);
-}
-
-TEST(MockSupportTest, customObjectWithFunctionComparator)
-{
-    MyTypeForTesting object(1);
-    MockFunctionComparator comparator(myTypeIsEqual, myTypeValueToString);
-    mock().installComparator("MyTypeForTesting", comparator);
-    mock().expectOneCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-    mock().actualCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-    mock().checkExpectations();
-    CHECK_NO_MOCK_FAILURE();
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, customObjectWithFunctionComparatorThatFailsCoversValueToString)
-{
-    MyTypeForTesting object(5);
-    MockFunctionComparator comparator(myTypeIsEqual, myTypeValueToString);
-    mock().installComparator("MyTypeForTesting", comparator);
-    expectations.addFunction("function")->withParameterOfType("MyTypeForTesting", "parameterName", &object);
-    MockExpectedCallsDidntHappenFailure failure(UtestShell::getCurrent(), expectations);
-    mock().expectOneCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-    mock().checkExpectations();
-    CHECK_EXPECTED_MOCK_FAILURE_LOCATION(failure, __FILE__, __LINE__);
-}
-
-TEST(MockSupportTest, customTypeOutputParameterSucceeds)
-{
-    MyTypeForTesting expectedObject(55);
-    MyTypeForTesting actualObject(99);
-    MyTypeForTestingCopier copier;
-    mock().installCopier("MyTypeForTesting", copier);
-
-    mock().expectOneCall("function").withOutputParameterOfTypeReturning("MyTypeForTesting", "parameterName", &expectedObject);
-    mock().actualCall("function").withOutputParameterOfType("MyTypeForTesting", "parameterName", &actualObject);
-    mock().checkExpectations();
-
-    CHECK_EQUAL(55, *(expectedObject.value));
-    CHECK_EQUAL(55, *(actualObject.value));
-    CHECK_NO_MOCK_FAILURE();
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, noActualCallForCustomTypeOutputParameter)
-{
-    MyTypeForTesting expectedObject(1);
-    MyTypeForTestingCopier copier;
-    mock().installCopier("MyTypeForTesting", copier);
-
-    expectations.addFunction("foo")->withOutputParameterOfTypeReturning("MyTypeForTesting", "output", &expectedObject);
-    MockExpectedCallsDidntHappenFailure expectedFailure(mockFailureTest(), expectations);
-
-    mock().expectOneCall("foo").withOutputParameterOfTypeReturning("MyTypeForTesting", "output", &expectedObject);
-    mock().checkExpectations();
-
-    CHECK_EXPECTED_MOCK_FAILURE(expectedFailure);
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, unexpectedCustomTypeOutputParameter)
-{
-    MyTypeForTesting actualObject(8834);
-    MyTypeForTestingCopier copier;
-    mock().installCopier("MyTypeForTesting", copier);
-
-    expectations.addFunction("foo");
-    MockNamedValue parameter("parameterName");
-    parameter.setObjectPointer("MyTypeForTesting", &actualObject);
-    MockUnexpectedOutputParameterFailure expectedFailure(mockFailureTest(), "foo", parameter, expectations);
-
-    mock().expectOneCall("foo");
-    mock().actualCall("foo").withOutputParameterOfType("MyTypeForTesting", "parameterName", &actualObject);
-    mock().checkExpectations();
-
-    CHECK_EXPECTED_MOCK_FAILURE(expectedFailure);
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, customTypeOutputParameterMissing)
-{
-    MyTypeForTesting expectedObject(123464);
-    MyTypeForTestingCopier copier;
-    mock().installCopier("MyTypeForTesting", copier);
-
-    expectations.addFunction("foo")->withOutputParameterOfTypeReturning("MyTypeForTesting", "output", &expectedObject);
-    MockExpectedParameterDidntHappenFailure expectedFailure(mockFailureTest(), "foo", expectations);
-
-    mock().expectOneCall("foo").withOutputParameterOfTypeReturning("MyTypeForTesting", "output", &expectedObject);
-    mock().actualCall("foo");
-    mock().checkExpectations();
-
-    CHECK_EXPECTED_MOCK_FAILURE(expectedFailure);
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, customTypeOutputParameterOfWrongType)
-{
-    MyTypeForTesting expectedObject(123464);
-    MyTypeForTesting actualObject(75646);
-    MyTypeForTestingCopier copier;
-    mock().installCopier("MyTypeForTesting", copier);
-
-    expectations.addFunction("foo")->withOutputParameterOfTypeReturning("MyTypeForTesting", "output", &expectedObject);
-    MockNamedValue parameter("output");
-    parameter.setObjectPointer("OtherTypeForTesting", &actualObject);
-    MockUnexpectedOutputParameterFailure expectedFailure(mockFailureTest(), "foo", parameter, expectations);
-
-    mock().expectOneCall("foo").withOutputParameterOfTypeReturning("MyTypeForTesting", "output", &expectedObject);
-    mock().actualCall("foo").withOutputParameterOfType("OtherTypeForTesting", "output", &actualObject);
-    mock().checkExpectations();
-
-    CHECK_EXPECTED_MOCK_FAILURE(expectedFailure);
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, noCopierForCustomTypeOutputParameter)
-{
-    MyTypeForTesting expectedObject(123464);
-    MyTypeForTesting actualObject(8834);
-
-    expectations.addFunction("foo")->withOutputParameterOfTypeReturning("MyTypeForTesting", "output", &expectedObject);
-    MockNoWayToCopyCustomTypeFailure expectedFailure(mockFailureTest(), "MyTypeForTesting");
-
-    mock().expectOneCall("foo").withOutputParameterOfTypeReturning("MyTypeForTesting", "output", &expectedObject);
-    mock().actualCall("foo").withOutputParameterOfType("MyTypeForTesting", "output", &actualObject);
-    mock().checkExpectations();
-
-    CHECK_EXPECTED_MOCK_FAILURE(expectedFailure);
-}
-
-TEST(MockSupportTest, twoCustomTypeOutputParameters)
-{
-    MyTypeForTesting expectedObject1(545);
-    MyTypeForTesting actualObject1(979);
-    MyTypeForTesting expectedObject2(123);
-    MyTypeForTesting actualObject2(4567);
-    MyTypeForTestingCopier copier;
-    mock().installCopier("MyTypeForTesting", copier);
-
-    mock().expectOneCall("function").withOutputParameterOfTypeReturning("MyTypeForTesting", "parameterName", &expectedObject1).withParameter("id", 1);
-    mock().expectOneCall("function").withOutputParameterOfTypeReturning("MyTypeForTesting", "parameterName", &expectedObject2).withParameter("id", 2);
-    mock().actualCall("function").withOutputParameterOfType("MyTypeForTesting", "parameterName", &actualObject1).withParameter("id", 1);
-    mock().actualCall("function").withOutputParameterOfType("MyTypeForTesting", "parameterName", &actualObject2).withParameter("id", 2);
-    mock().checkExpectations();
-
-    CHECK_EQUAL(545, *(expectedObject1.value));
-    CHECK_EQUAL(545, *(actualObject1.value));
-    CHECK_EQUAL(123, *(expectedObject2.value));
-    CHECK_EQUAL(123, *(actualObject2.value));
-    CHECK_NO_MOCK_FAILURE();
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, twoInterleavedCustomTypeOutputParameters)
-{
-    MyTypeForTesting expectedObject1(9545);
-    MyTypeForTesting actualObject1(79);
-    MyTypeForTesting expectedObject2(132);
-    MyTypeForTesting actualObject2(743);
-    MyTypeForTestingCopier copier;
-
-    mock().installCopier("MyTypeForTesting", copier);
-    mock().expectOneCall("function").withOutputParameterOfTypeReturning("MyTypeForTesting", "parameterName", &expectedObject1).withParameter("id", 1);
-    mock().expectOneCall("function").withOutputParameterOfTypeReturning("MyTypeForTesting", "parameterName", &expectedObject2).withParameter("id", 2);
-    mock().actualCall("function").withOutputParameterOfType("MyTypeForTesting", "parameterName", &actualObject2).withParameter("id", 2);
-    mock().actualCall("function").withOutputParameterOfType("MyTypeForTesting", "parameterName", &actualObject1).withParameter("id", 1);
-    mock().checkExpectations();
-
-    CHECK_EQUAL(9545, *(expectedObject1.value));
-    CHECK_EQUAL(9545, *(actualObject1.value));
-    CHECK_EQUAL(132, *(expectedObject2.value));
-    CHECK_EQUAL(132, *(actualObject2.value));
-    CHECK_NO_MOCK_FAILURE();
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, twoDifferentCustomTypeOutputParametersInSameFunctionCallSucceeds)
-{
-    MyTypeForTesting expectedObject1(11);
-    MyTypeForTesting actualObject1(22);
-    MyTypeForTesting expectedObject2(33);
-    MyTypeForTesting actualObject2(44);
-    MyTypeForTestingCopier copier;
-    mock().installCopier("MyTypeForTesting", copier);
-
-    mock().expectOneCall("foo")
-        .withOutputParameterOfTypeReturning("MyTypeForTesting", "bar", &expectedObject1)
-        .withOutputParameterOfTypeReturning("MyTypeForTesting", "foobar", &expectedObject2);
-    mock().actualCall("foo")
-        .withOutputParameterOfType("MyTypeForTesting", "bar", &actualObject1)
-        .withOutputParameterOfType("MyTypeForTesting", "foobar", &actualObject2);
-    mock().checkExpectations();
-
-    CHECK_EQUAL(11, *(expectedObject1.value));
-    CHECK_EQUAL(11, *(actualObject1.value));
-    CHECK_EQUAL(33, *(expectedObject2.value));
-    CHECK_EQUAL(33, *(actualObject2.value));
-    CHECK_NO_MOCK_FAILURE();
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, customTypeOutputAndInputParametersOfSameNameInDifferentFunctionCallsOfSameFunctionSucceeds)
-{
-    MyTypeForTesting expectedObject1(911);
-    MyTypeForTesting actualObject1(6576878);
-    MyTypeForTesting expectedObject2(123);
-    MyTypeForTesting actualObject2(123);
-    MyTypeForTestingCopier copier;
-    MyTypeForTestingComparator comparator;
-    mock().installCopier("MyTypeForTesting", copier);
-    mock().installComparator("MyTypeForTesting", comparator);
-
-    mock().expectOneCall("foo").withOutputParameterOfTypeReturning("MyTypeForTesting", "bar", &expectedObject1);
-    mock().expectOneCall("foo").withParameterOfType("MyTypeForTesting", "bar", &expectedObject2);
-    mock().actualCall("foo").withOutputParameterOfType("MyTypeForTesting", "bar", &actualObject1);
-    mock().actualCall("foo").withParameterOfType("MyTypeForTesting", "bar", &actualObject2);
-    mock().checkExpectations();
-
-    CHECK_EQUAL(911, *(expectedObject1.value));
-    CHECK_EQUAL(911, *(actualObject1.value));
-    CHECK_EQUAL(123, *(expectedObject2.value));
-    CHECK_EQUAL(123, *(actualObject2.value));
-    CHECK_NO_MOCK_FAILURE();
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, twoCustomTypeOutputParametersOfSameNameInDifferentFunctionsSucceeds)
-{
-    MyTypeForTesting expectedObject1(657);
-    MyTypeForTesting actualObject1(984465);
-    MyTypeForTesting expectedObject2(987);
-    MyTypeForTesting actualObject2(987);
-    MyTypeForTestingCopier copier;
-    MyTypeForTestingComparator comparator;
-    mock().installCopier("MyTypeForTesting", copier);
-    mock().installComparator("MyTypeForTesting", comparator);
-
-    mock().expectOneCall("foo1").withOutputParameterOfTypeReturning("MyTypeForTesting", "bar", &expectedObject1);
-    mock().expectOneCall("foo2").withParameterOfType("MyTypeForTesting", "bar", &expectedObject2);
-    mock().actualCall("foo1").withOutputParameterOfType("MyTypeForTesting", "bar", &actualObject1);
-    mock().actualCall("foo2").withParameterOfType("MyTypeForTesting", "bar", &actualObject2);
-    mock().checkExpectations();
-
-    CHECK_EQUAL(657, *(expectedObject1.value));
-    CHECK_EQUAL(657, *(actualObject1.value));
-    CHECK_EQUAL(987, *(expectedObject2.value));
-    CHECK_EQUAL(987, *(actualObject2.value));
-    CHECK_NO_MOCK_FAILURE();
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, customTypeOutputAndInputParameterOfSameTypeInSameFunctionCall)
-{
-    MyTypeForTesting expectedObject1(45);
-    MyTypeForTesting actualObject1(45);
-    MyTypeForTesting expectedObject2(987765443);
-    MyTypeForTesting actualObject2(0);
-    MyTypeForTestingCopier copier;
-    MyTypeForTestingComparator comparator;
-    mock().installCopier("MyTypeForTesting", copier);
-    mock().installComparator("MyTypeForTesting", comparator);
-
-    mock().expectOneCall("foo")
-          .withParameterOfType("MyTypeForTesting", "bar", &expectedObject1)
-          .withOutputParameterOfTypeReturning("MyTypeForTesting", "bar", &expectedObject2);
-    mock().actualCall("foo")
-          .withParameterOfType("MyTypeForTesting", "bar", &actualObject1)
-          .withOutputParameterOfType("MyTypeForTesting", "bar", &actualObject2);
-    mock().checkExpectations();
-
-    CHECK_EQUAL(45, *(expectedObject1.value));
-    CHECK_EQUAL(45, *(actualObject1.value));
-    CHECK_EQUAL(987765443, *(expectedObject2.value));
-    CHECK_EQUAL(987765443, *(actualObject2.value));
-    CHECK_NO_MOCK_FAILURE();
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, customTypeOutputParameterTraced)
-{
-    MyTypeForTesting actualObject(676789);
-    MyTypeForTestingCopier copier;
-    mock().installCopier("MyTypeForTesting", copier);
-    mock().tracing(true);
-
-    mock().actualCall("someFunc").withOutputParameterOfType("MyTypeForTesting", "someParameter", &actualObject);
-    mock().checkExpectations();
-
-    STRCMP_CONTAINS("Function name:someFunc MyTypeForTesting someParameter:", mock().getTraceOutput());
-    CHECK_NO_MOCK_FAILURE();
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, customTypeOutputParameterWithIgnoredParameters)
-{
-    MyTypeForTesting expectedObject(444537909);
-    MyTypeForTesting actualObject(98765);
-    MyTypeForTestingCopier copier;
-    mock().installCopier("MyTypeForTesting", copier);
-
-    mock().expectOneCall("foo").withOutputParameterOfTypeReturning("MyTypeForTesting", "bar", &expectedObject).ignoreOtherParameters();
-    mock().actualCall("foo").withOutputParameterOfType("MyTypeForTesting", "bar", &actualObject).withParameter("other", 1);
-    mock().checkExpectations();
-
-    CHECK_EQUAL(444537909, *(expectedObject.value));
-    CHECK_EQUAL(444537909, *(actualObject.value));
-    CHECK_NO_MOCK_FAILURE();
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
-static void myTypeCopy(void* dst_, const void* src_)
-{
-    MyTypeForTesting* dst = (MyTypeForTesting*) dst_;
-    const MyTypeForTesting* src = (const MyTypeForTesting*) src_;
-    *(dst->value) = *(src->value);
-}
-
-TEST(MockSupportTest, customObjectWithFunctionCopier)
-{
-    MyTypeForTesting expectedObject(9874452);
-    MyTypeForTesting actualObject(2034);
-    MockFunctionCopier copier(myTypeCopy);
-    mock().installCopier("MyTypeForTesting", copier);
-
-    mock().expectOneCall("function").withOutputParameterOfTypeReturning("MyTypeForTesting", "parameterName", &expectedObject);
-    mock().actualCall("function").withOutputParameterOfType("MyTypeForTesting", "parameterName", &actualObject);
-    mock().checkExpectations();
-
-    CHECK_EQUAL(9874452, *(expectedObject.value));
-    CHECK_EQUAL(9874452, *(actualObject.value));
-    CHECK_NO_MOCK_FAILURE();
-
-    mock().removeAllComparatorsAndCopiers();
-}
-
 
 TEST(MockSupportTest, disableEnable)
 {
@@ -866,54 +436,6 @@ TEST(MockSupportTest, reporterIsInheritedInHierarchicalMocks)
     mock("differentScope").actualCall("foobar");
 
     MockUnexpectedCallHappenedFailure expectedFailure(mockFailureTest(), "foobar", expectations);
-    CHECK_EXPECTED_MOCK_FAILURE(expectedFailure);
-}
-
-TEST(MockSupportTest, installComparatorWorksHierarchicalOnBothExistingAndDynamicallyCreatedMockSupports)
-{
-    MyTypeForTesting object(1);
-    MyTypeForTestingComparator comparator;
-
-    mock("existing");
-    mock().installComparator("MyTypeForTesting", comparator);
-    mock("existing").expectOneCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-    mock("existing").actualCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-    mock("dynamic").expectOneCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-    mock("dynamic").actualCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-
-    mock().checkExpectations();
-    CHECK_NO_MOCK_FAILURE();
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, installComparatorsWorksHierarchical)
-{
-    MyTypeForTesting object(1);
-    MyTypeForTestingComparator comparator;
-    MockNamedValueComparatorsAndCopiersRepository repos;
-    repos.installComparator("MyTypeForTesting", comparator);
-
-    mock("existing");
-    mock().installComparatorsAndCopiers(repos);
-    mock("existing").expectOneCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-    mock("existing").actualCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-
-    mock().checkExpectations();
-    CHECK_NO_MOCK_FAILURE();
-    mock().removeAllComparatorsAndCopiers();
-}
-
-TEST(MockSupportTest, removeComparatorsAndCopiersWorksHierachically)
-{
-    MyTypeForTesting object(1);
-    MyTypeForTestingComparator comparator;
-
-    mock("scope").installComparator("MyTypeForTesting", comparator);
-    mock().removeAllComparatorsAndCopiers();
-    mock("scope").expectOneCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-    mock("scope").actualCall("function").withParameterOfType("MyTypeForTesting", "parameterName", &object);
-
-    MockNoWayToCompareCustomTypeFailure expectedFailure(mockFailureTest(), "MyTypeForTesting");
     CHECK_EXPECTED_MOCK_FAILURE(expectedFailure);
 }
